+++ fda-metadata-db/entities/src/main/java/at/tuwien/entities/database/Database.java
@@ -76,7 +76,7 @@ public class Database {
     private String publisher;
 
     @Column
-    private Short publicationYear;
+    private Integer publicationYear;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.MERGE)
     @JoinColumns({
